// src/app/modules/card-panel/component-loader/standard-components/standard-types.ts
export enum StandardComponentTypes {
  /**
   * This is a basic text header.
   * Inputs: textToShow - This is the string to show in the header.
   */
  TextHeader = 'TextHeader',

  /**
   * This is a basic text body.
   * Inputs: textToShow - This is the string to show in the header.
   */
  TextBody = 'TextBody',
}